requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.37
botocore==1.42.37
click==8.3.1
cmdstanpy==1.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.128.0
fonttools==4.61.1
h11==0.16.0
holidays==0.89
idna==3.11
importlib_resources==6.5.2
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
prophet==1.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
stanio==0.5.1
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
